services/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
pymysql>=1.1,<2.0
redis>=5.0,<6.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
cryptography>=44.0,<45.0
